Removed the `Option` namespace
