+++ include/options/business/nav-menus.php
@@ -3,9 +3,9 @@
  * @package Polylang
  */
 
-namespace WP_Syntex\Polylang\Options\Option\Business;
+namespace WP_Syntex\Polylang\Options\Business;
 
-use WP_Syntex\Polylang\Options\Option\Primitive\Map;
+use WP_Syntex\Polylang\Options\Primitive\Map;
 
 defined( 'ABSPATH' ) || exit;
 
@@ -14,7 +14,7 @@ defined( 'ABSPATH' ) || exit;
  *
  * @since 3.7
  *
- * @phpstan-import-type SchemaType from \WP_Syntex\Polylang\Options\Option\Abstract_Option
+ * @phpstan-import-type SchemaType from \WP_Syntex\Polylang\Options\Abstract_Option
  */
 class Nav_Menu extends Map {
 	/**
